+++ src/api.h
@@ -15,13 +15,13 @@ int
 crypto_sign_keypair(unsigned char *pk, unsigned char *sk);
 
 int
-crypto_sign(unsigned char *sm, unsigned long long *smlen,
-            const unsigned char *m, unsigned long long mlen,
+crypto_sign(unsigned char *sm, uint64_t *smlen,
+            const unsigned char *m, uint64_t mlen,
             const unsigned char *sk);
 
 int
-crypto_sign_open(unsigned char *m, unsigned long long *mlen,
-                 const unsigned char *sm, unsigned long long smlen,
+crypto_sign_open(unsigned char *m, uint64_t *mlen,
+                 const unsigned char *sm, uint64_t smlen,
                  const unsigned char *pk);
 
 #endif /* api_h */
